Use `IGNORE` by default in creation methods
Conflicts in things like the ID should be automatically ignored, other
possible collisions like a collection's name is left to the dispatcher
to figure out.

For example, in the case of a duplicate collection (name), we can just
add a number, but if the event specifies a specific ID that already
exists, we don't care what the name is, we need to ignore it.